+++ experimental/SymmetricPolynomials/src/SymmetricPolynomials.jl
@@ -1,0 +1,21 @@
+module SymmetricPolynomials
+
+using Oscar
+
+# Export the main functions
+export schur, grothendieck, dual_grothendieck
+export p, e, h
+export schur_expansion, grothendieck_expansion, dual_grothendieck_expansion
+export pieri_rule, dual_pieri_rule, murnaghan_nakayama_rule
+export G_pieri_rule, dual_G_pieri_rule, G_murnaghan_nakayama_rule
+export g_pieri_rule, dual_g_pieri_rule, g_murnaghan_nakayama_rule
+export littlewood_richardson_rule, G_littlewood_richardson_rule, g_littlewood_richardson_rule
+
+include("schur.jl")
+include("symmetric_functions.jl")
+include("grothendieck_polynomials.jl")
+include("dual_grothendieck_polynomials.jl")
+include("expansion_functions.jl")
+include("rule_functions.jl")
+
+end
